fix: Add Maude dependencies
Judith uses the front part of Maude. However, in Maude
the front part depends on the back part. And that back part
is unknown to Judith. So it can't/won't draft it prior to
drafting the Maude front which will cause the Maude front
part to fail.

This commit addresses that.